src/test/java/za/com/anyCompany/OrderServiceTest.java
package za.com.anyCompany;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import za.com.anyCompany.DTO.RequestDto;
import za.com.anyCompany.DTO.ResponseDto;
import za.com.anyCompany.Repo.OrderRepo;
import za.com.anyCompany.Repo.ProductRepo;
import za.com.anyCompany.Service.AddressService;
import za.com.anyCompany.Service.CustomerService;
import za.com.anyCompany.Service.OrderService;
import za.com.anyCompany.model.AddressModel;
import za.com.anyCompany.model.CustomerModel;
import za.com.anyCompany.model.ProductModel;

import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;

public class OrderServiceTest {

    @Mock
    private ProductRepo productRepo;

    @Mock
    private OrderRepo orderRepo;

    @Mock
    private CustomerService customerService;

    @Mock
    private AddressService addressService;

    @InjectMocks
    private OrderService orderService;

    @BeforeEach
    public void setUp() {
        MockitoAnnotations.initMocks(this);
    }

    @Test
    public void testPlaceOrder_WithValidRequest_ShouldReturnSuccess() {
        // Mock data
        RequestDto requestDto = new RequestDto();
        // Set the necessary data in the requestDto

        // Mock the behavior of the dependencies
        when(productRepo.findByProductName(any(String.class))).thenReturn(new ProductModel());
        when(addressService.findShippingAddress(any(CustomerModel.class))).thenReturn(new AddressModel());

        // Call the method under test
        ResponseDto responseDto = orderService.placeOrder(requestDto);

        // Assertions
        Assertions.assertTrue(responseDto.isStatus());
        Assertions.assertEquals("Order has been successfully processed", responseDto.getMessage());
        // Add more assertions as needed
    }

    // Add more test methods to cover different scenarios and edge cases

}
